+++ src/PhyloAcc-interface/phyloacc_lib/templates.py
@@ -18,6 +18,7 @@ OUTGROUP {outgroup}
 CONSERVE {conserved}
 NUM_THREAD {procs_per_job}
 {phyloacc_opts}
+{dollo_str}
 """
 
     return phyloacc_template;
@@ -27,7 +28,7 @@ NUM_THREAD {procs_per_job}
 def snakemake():
 
     smk_template = """#############################################################################
-# Pipeline for running phyloacc per locus
+# Pipeline for running phyloacc batches
 # Generated from: {cmd}
 # On: {dt}
 #############################################################################
@@ -51,7 +52,7 @@ GT_BATCHES = config["gt_batch_list"];
 #############################################################################
 
 iqtree_loci = [];
-{run_char}iqtree_loci = [ f.replace(".fa", "") for f in os.listdir(os.path.join(IQTREEDIR, "alns")) if f.endswith(".fa") ];
+{theta_char}iqtree_loci = [ f.replace(".fa", "") for f in os.listdir(os.path.join(IQTREEDIR, "alns")) if f.endswith(".fa") ];
 # Reads the alignments to make trees from if --theta is specified
 
 #############################################################################
@@ -62,9 +63,9 @@ localrules: all
 
 rule all:
     input:
-        {run_char}expand(os.path.join(IQTREEDIR, "gene-trees", "{{locus}}", "{{locus}}.treefile"), locus=iqtree_loci),
-        {run_char}os.path.join(IQTREEDIR, "gene-trees", "all-gene-trees.treefile"),
-        {run_char}"{coal_tree_path}",
+        {theta_char}expand(os.path.join(IQTREEDIR, "gene-trees", "{{locus}}", "{{locus}}.treefile"), locus=iqtree_loci),
+        {theta_char}os.path.join(IQTREEDIR, "gene-trees", "all-gene-trees.treefile"),
+        {theta_char}"{coal_tree_path}",
         {phyloacc_char}expand(os.path.join(OUTDIR, "{{st_batch}}-phyloacc-st-out", "{{st_batch}}_elem_lik.txt"), st_batch=ST_BATCHES),
         {phyloacc_char}expand(os.path.join(OUTDIR, "{{gt_batch}}-phyloacc-gt-out", "{{gt_batch}}_elem_lik.txt"), gt_batch=GT_BATCHES)
 # This rule checks for expected outputs from the rules